refactor(coeus-ops): Split activation module tree
Keep the public activation exports stable while assigning softmax and GLU to named leaf modules. Record the boundary in ADR-0021.
